Added support for output columns of type SQLT_TIMESTAMP or SQLT_TIMESTAMP_TZ
This change fixes a bug, where SQL queries of the form

     SELECT TO_TIMESTAMP(sysdate) FROM dual

lead to errors for the form:

     Database operation "getrow" failed (exception 1406, "nsoracle.c:3659:Ns_OracleGetRow:
     error in `OCIStmtFetch ()': ORA-01406: fetched column value was truncated

The driver needs for several output types special rules, where the
timestamp cases were not supported so far. It is also recommended
to set the according environment variables specifying the output
format in the configuration server of NaviServer, such as e.g.

    set ::env(NLS_TIMESTAMP_FORMAT) "YYYY-MM-DD HH24:MI:SS.FF6"
    set ::env(NLS_TIMESTAMP_TZ_FORMAT) "YYYY-MM-DD HH24:MI:SS.FF6 TZH:TZM"

For testing in you local Oracle installation, you might test the output
formats (and the required sizes with the following snippet for sqlplus:

   COLUMN localtimestamp format a40
   COLUMN systimestamp format a40
   COLUMN ts_bytes format a80

   alter session set nls_timestamp_format='YYYY-MM-DD HH24:MI:SS.FF6';
   select localtimestamp, length(localtimestamp), dump(localtimestamp) ts_bytes from dual;

   alter session set nls_timestamp_tz_format='YYYY-MM-DD HH24:MI:SS.FF6 TZH:TZM';
   select systimestamp, length(systimestamp), dump(systimestamp) ts_bytes from dual;

   alter session set nls_timestamp_tz_format='YYYY-MM-DD HH24:MI:SS.FF6 TZR';
   select systimestamp, length(systimestamp), dump(systimestamp) ts_bytes from dual;