Added experimental boolean configuration parameter "ConvertEncoding"
When this Boolean parameter is true, the connection is created with
OCIEnvNlsCreate() instead of OCIEnvCreate(). OCIEnvNlsCreate was
introduced with Oracle9 and is an enhanced version of OCIEnvCreate()
which is used by default (legacy setting) or when ConvertEncoding is
not true.

According to the documentation, using OCIEnvCreate() is equivalent to
OCIEnvNlsCreate() with character IDs set to 0. In this case, the
NLS_LANG setting is used, which should be "en_US.UTF-8" or
similar. However, since Oracle supports multiple UTF variants, it
sounds less adventurous to provide the encoding, which Tcl expects
(AL32UTF8: 1 to 4 bytes UTF-8), explicitly.